Move call_model to base agent
