+++ src/jutulgpt/agents/autonomous_agent.py
@@ -3,7 +3,6 @@ from __future__ import annotations
 from typing import Any, Callable, Optional, Sequence, Union
 
 from langchain_core.language_models import LanguageModelLike
-from langchain_core.messages import AIMessage
 from langchain_core.runnables import RunnableConfig
 from langchain_core.tools import BaseTool
 from langgraph.graph import END, StateGraph
@@ -11,7 +10,7 @@ from langgraph.prebuilt import ToolNode
 
 from jutulgpt.agents.agent_base import BaseAgent
 from jutulgpt.configuration import BaseConfiguration, cli_mode, mcp_mode
-from jutulgpt.state import MCPInputState, MCPOutputState, State
+from jutulgpt.state import MCPInputState, MCPOutputState
 from jutulgpt.tools import (
     execute_terminal_command,
     get_working_directory,
@@ -24,8 +23,6 @@ from jutulgpt.tools import (
     run_julia_linter,
     write_to_file,
 )
-from jutulgpt.utils.code_parsing import get_code_from_response
-from jutulgpt.utils.model import get_message_text
 
 
 class AutonomousAgent(BaseAgent):
@@ -111,33 +108,6 @@ class AutonomousAgent(BaseAgent):
         """
         configuration = BaseConfiguration.from_runnable_config(config)
         return configuration.autonomous_agent_prompt
-
-    def call_model(self, state: State, config: RunnableConfig) -> dict:
-        """Call the model with the current state."""
-
-        response = self.invoke_model(state=state, config=config)
-
-        # Check if we need more steps
-        if self._are_more_steps_needed(state, response):
-            fallback = AIMessage(
-                id=response.id,
-                content="Sorry, need more steps to process this request.",
-            )
-            return {"messages": self._finalize_context(fallback)}
-
-        # With OpenAI Responses API, response.content may be a list of content blocks.
-        # Always use the normalized text view for downstream parsing.
-        response_text = get_message_text(response)
-        code_block = get_code_from_response(response=response_text)
-
-        # Finalize context: bundle response with any pending state changes
-        messages = self._finalize_context(response)
-        return {
-            "messages": messages,
-            "code_block": code_block,
-            "error": False,
-            "mcp_answer": response_text,
-        }
 
 
 autonomous_agent = AutonomousAgent(
